(PC-23190)[API] chore: extract patch event test
